Test updates (#239)
* Move tests to test-directory
* Update npm test scripts
