+++ .gitignore
@@ -3,5 +3,4 @@ node_modules
 coverage
 dist
 .nyc_output
-test/
 .env
