+++ test-fixtures/index/20/metadata.json
@@ -16,7 +16,7 @@
       }
     },
     "maxMatches": 1,
-    "maxCandidates": 1
+    "maxCandidates": 100
   },
   "expectedMatchStatus": true,
   "expectedStopReason": "maxMatches",
@@ -24,7 +24,7 @@
   "requests": [
     {
       "method": "get",
-      "url": "/?operation=searchRetrieve&query=dc.identifier%3D978-952-62-2477-0&startRecord=1&version=2.0&maximumRecords=1",
+      "url": "/?operation=searchRetrieve&query=dc.identifier%3D978-952-62-2477-0&startRecord=1&version=2.0&maximumRecords=3",
       "status": 200
     }
   ],
