+++ test/candidate-search/index.test.js
@@ -4,10 +4,12 @@ import {READERS} from '@natlibfi/fixura';
 import generateTests from '@natlibfi/fixugen-http-client';
 import {MarcRecord} from '@natlibfi/marc-record';
 import {Error as MatchingError} from '@natlibfi/melinda-commons';
-import createSearchInterface, {CandidateSearchError} from './index.js';
+import createSearchInterface, {CandidateSearchError} from '../../src/candidate-search/index.js';
 import createDebugLogger from 'debug';
 
 const debug = createDebugLogger('@natlibfi/melinda-record-matching:candidate-search:test');
+
+// NOTE: it seems that "only" does not work when using generateTests from @natlibfi/fixugen-http-client'
 
 describe('candidate-search', () => {
   generateTests({
